requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cloudscraper==1.2.71
exceptiongroup==1.2.2
flake8==6.0.0
flake8-isort==6.0.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
isort==5.13.2
lxml==5.2.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
outcome==1.3.0.post0
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.0.1
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
